Use io.Reader and io.Writer interfaces
This refactor jsonlogic functions to get the logic and data as an
io.Reader interface and the result must be an io.Writer interface
where the output of the logic execution will be writed.

This breaks compatibility with the old signature where the
input and output is based on `interface{}`. I do think this BC makes
sense and we achieve a better code using those interfaces.